+++ src/algorithms/local_search/local_search.cpp
@@ -25,6 +25,7 @@ All rights reserved (see LICENSE).
 #include "problems/cvrp/operators/reverse_two_opt.h"
 #include "problems/cvrp/operators/route_exchange.h"
 #include "problems/cvrp/operators/two_opt.h"
+#include "problems/cvrp/operators/unassigned_exchange.h"
 #include "problems/vrptw/operators/cross_exchange.h"
 #include "problems/vrptw/operators/exchange.h"
 #include "problems/vrptw/operators/intra_cross_exchange.h"
@@ -39,12 +40,14 @@ All rights reserved (see LICENSE).
 #include "problems/vrptw/operators/reverse_two_opt.h"
 #include "problems/vrptw/operators/route_exchange.h"
 #include "problems/vrptw/operators/two_opt.h"
+#include "problems/vrptw/operators/unassigned_exchange.h"
 #include "utils/helpers.h"
 
 namespace vroom {
 namespace ls {
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -60,6 +63,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 LocalSearch<Route,
+            UnassignedExchange,
             Exchange,
             CrossExchange,
             MixedExchange,
@@ -110,6 +114,7 @@ LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -125,6 +130,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 void LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -157,12 +163,12 @@ void LocalSearch<Route,
     Index best_delivery_r = 0;
 
     for (const auto j : _sol_state.unassigned) {
-      auto& current_job = _input.jobs[j];
+      const auto& current_job = _input.jobs[j];
       if (current_job.type == JOB_TYPE::DELIVERY) {
         continue;
       }
 
-      auto job_priority = current_job.priority;
+      const auto job_priority = current_job.priority;
 
       if (job_priority < best_priority) {
         // Insert higher priority jobs first.
@@ -176,7 +182,7 @@ void LocalSearch<Route,
 
       if (current_job.type == JOB_TYPE::SINGLE) {
         for (std::size_t i = 0; i < routes.size(); ++i) {
-          auto v = routes[i];
+          const auto v = routes[i];
           const auto& v_target = _input.vehicles[v];
 
           if (!_input.vehicle_ok_with_job(v, j)) {
@@ -205,7 +211,7 @@ void LocalSearch<Route,
 
       if (current_job.type == JOB_TYPE::PICKUP) {
         for (std::size_t i = 0; i < routes.size(); ++i) {
-          auto v = routes[i];
+          const auto v = routes[i];
           const auto& v_target = _input.vehicles[v];
 
           if (!_input.vehicle_ok_with_job(v, j)) {
@@ -334,14 +340,15 @@ void LocalSearch<Route,
       // Find best route for current job based on cost of addition and
       // regret cost of not adding.
       for (std::size_t i = 0; i < routes.size(); ++i) {
-        auto addition_cost = best_costs[i];
+        const auto addition_cost = best_costs[i];
         if (addition_cost == std::numeric_limits<Gain>::max()) {
           continue;
         }
-        Gain regret_cost = (i == smallest_idx) ? second_smallest : smallest;
+        const Gain regret_cost =
+          (i == smallest_idx) ? second_smallest : smallest;
 
-        double eval = static_cast<double>(addition_cost) -
-                      regret_coeff * static_cast<double>(regret_cost);
+        const double eval = static_cast<double>(addition_cost) -
+                            regret_coeff * static_cast<double>(regret_cost);
 
         if ((job_priority > best_priority) or
             (job_priority == best_priority and eval < best_cost)) {
@@ -364,7 +371,7 @@ void LocalSearch<Route,
 
     if (job_added) {
       _sol_state.unassigned.erase(best_job_rank);
-      auto& best_job = _input.jobs[best_job_rank];
+      const auto& best_job = _input.jobs[best_job_rank];
 
       if (best_job.type == JOB_TYPE::SINGLE) {
         _sol[best_route].add(_input, best_job_rank, best_rank);
@@ -397,6 +404,7 @@ void LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -412,6 +420,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 void LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -426,6 +435,7 @@ void LocalSearch<Route,
                  IntraOrOpt,
                  PDShift,
                  RouteExchange>::run_ls_step() {
+  // Store best move involving a pair of routes.
   std::vector<std::vector<std::unique_ptr<Operator>>> best_ops(_nb_vehicles);
   for (std::size_t v = 0; v < _nb_vehicles; ++v) {
     best_ops[v] = std::vector<std::unique_ptr<Operator>>(_nb_vehicles);
@@ -442,20 +452,87 @@ void LocalSearch<Route,
     }
   }
 
-  std::vector<std::vector<Gain>> best_gains(_nb_vehicles,
-                                            std::vector<Gain>(_nb_vehicles, 0));
+  // Store best gain for matching move.
+  std::vector<std::vector<Gain>>
+    best_gains(_nb_vehicles,
+               std::vector<Gain>(_nb_vehicles,
+                                 std::numeric_limits<Gain>::min()));
+
+  // Store best priority increase for matching move. Only operators
+  // involving a single route and unassigned jobs can change overall
+  // priority (currently only UnassignedExchange).
+  std::vector<Priority> best_priorities(_nb_vehicles, 0);
 
   Gain best_gain = 1;
+  Priority best_priority = 0;
 
-  while (best_gain > 0) {
+  while (best_gain > 0 or best_priority > 0) {
     // Operators applied to a pair of (different) routes.
 
     if (_input.has_jobs()) {
       // Move(s) that don't make sense for shipment-only instances.
 
+      // Unassigned-exchange stuff
+      for (const Index u : _sol_state.unassigned) {
+        if (_input.jobs[u].type != JOB_TYPE::SINGLE) {
+          continue;
+        }
+
+        Priority u_priority = _input.jobs[u].priority;
+        for (const auto& s_t : s_t_pairs) {
+          if (s_t.first != s_t.second or
+              !_input.vehicle_ok_with_job(s_t.first, u) or
+              _sol[s_t.first].empty()) {
+            continue;
+          }
+
+          for (unsigned s_rank = 0; s_rank < _sol[s_t.first].size(); ++s_rank) {
+            const auto& current_job =
+              _input.jobs[_sol[s_t.first].route[s_rank]];
+            if (current_job.type != JOB_TYPE::SINGLE or
+                u_priority < current_job.priority) {
+              continue;
+            }
+
+            const Priority priority_gain = u_priority - current_job.priority;
+
+            if (best_priorities[s_t.first] <= priority_gain) {
+              for (unsigned t_rank = 0; t_rank <= _sol[s_t.first].size();
+                   ++t_rank) {
+                if (t_rank == s_rank + 1) {
+                  // Same move as with t_rank == s_rank.
+                  continue;
+                }
+                UnassignedExchange r(_input,
+                                     _sol_state,
+                                     _sol_state.unassigned,
+                                     _sol[s_t.first],
+                                     s_t.first,
+                                     s_rank,
+                                     t_rank,
+                                     u);
+
+                bool better_if_valid =
+                  (best_priorities[s_t.first] < priority_gain) or
+                  (best_priorities[s_t.first] == priority_gain and
+                   r.gain() > best_gains[s_t.first][s_t.first]);
+
+                if (better_if_valid and r.is_valid()) {
+                  best_priorities[s_t.first] = priority_gain;
+                  best_gains[s_t.first][s_t.first] = r.gain();
+                  best_ops[s_t.first][s_t.first] =
+                    std::make_unique<UnassignedExchange>(r);
+                }
+              }
+            }
+          }
+        }
+      }
+
       // Exchange stuff
       for (const auto& s_t : s_t_pairs) {
         if (s_t.second <= s_t.first or // This operator is symmetric.
+            best_priorities[s_t.first] > 0 or best_priorities[s_t.second] > 0 or
             _sol[s_t.first].size() == 0 or _sol[s_t.second].size() == 0) {
           continue;
         }
@@ -500,6 +577,7 @@ void LocalSearch<Route,
     // CROSS-exchange stuff
     for (const auto& s_t : s_t_pairs) {
       if (s_t.second <= s_t.first or // This operator is symmetric.
+          best_priorities[s_t.first] > 0 or best_priorities[s_t.second] > 0 or
           _sol[s_t.first].size() < 2 or _sol[s_t.second].size() < 2) {
         continue;
       }
@@ -579,7 +657,8 @@ void LocalSearch<Route,
     if (_input.has_jobs()) {
       // Mixed-exchange stuff
       for (const auto& s_t : s_t_pairs) {
-        if (s_t.first == s_t.second or _sol[s_t.first].size() == 0 or
+        if (s_t.first == s_t.second or best_priorities[s_t.first] > 0 or
+            best_priorities[s_t.second] > 0 or _sol[s_t.first].size() == 0 or
             _sol[s_t.second].size() < 2) {
           continue;
         }
@@ -644,8 +723,8 @@ void LocalSearch<Route,
 
     // 2-opt* stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.second <= s_t.first) {
-        // This operator is symmetric.
+      if (s_t.second <= s_t.first or // This operator is symmetric.
+          best_priorities[s_t.first] > 0 or best_priorities[s_t.second] > 0) {
         continue;
       }
 
@@ -696,7 +775,8 @@ void LocalSearch<Route,
 
     // Reverse 2-opt* stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first == s_t.second) {
+      if (s_t.first == s_t.second or best_priorities[s_t.first] > 0 or
+          best_priorities[s_t.second] > 0) {
         continue;
       }
 
@@ -745,8 +825,8 @@ void LocalSearch<Route,
 
       // Relocate stuff
       for (const auto& s_t : s_t_pairs) {
-        if (s_t.first == s_t.second or _sol[s_t.first].size() == 0) {
-          // Don't try to put things from an empty vehicle.
+        if (s_t.first == s_t.second or best_priorities[s_t.first] > 0 or
+            best_priorities[s_t.second] > 0 or _sol[s_t.first].size() == 0) {
           continue;
         }
 
@@ -787,8 +867,8 @@ void LocalSearch<Route,
 
       // Or-opt stuff
       for (const auto& s_t : s_t_pairs) {
-        if (s_t.first == s_t.second or _sol[s_t.first].size() < 2) {
-          // Don't try to move things from a (near-)empty vehicle.
+        if (s_t.first == s_t.second or best_priorities[s_t.first] > 0 or
+            best_priorities[s_t.second] > 0 or _sol[s_t.first].size() < 2) {
           continue;
         }
 
@@ -844,7 +924,8 @@ void LocalSearch<Route,
 
     // Intra exchange stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first != s_t.second or _sol[s_t.first].size() < 3) {
+      if (s_t.first != s_t.second or best_priorities[s_t.first] > 0 or
+          _sol[s_t.first].size() < 3) {
         continue;
       }
 
@@ -881,7 +962,8 @@ void LocalSearch<Route,
 
     // Intra CROSS-exchange stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first != s_t.second or _sol[s_t.first].size() < 5) {
+      if (s_t.first != s_t.second or best_priorities[s_t.first] > 0 or
+          _sol[s_t.first].size() < 5) {
         continue;
       }
 
@@ -944,7 +1026,8 @@ void LocalSearch<Route,
 
     // Intra mixed-exchange stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first != s_t.second or _sol[s_t.first].size() < 4) {
+      if (s_t.first != s_t.second or best_priorities[s_t.first] > 0 or
+          _sol[s_t.first].size() < 4) {
         continue;
       }
 
@@ -997,7 +1080,8 @@ void LocalSearch<Route,
 
     // Intra relocate stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first != s_t.second or _sol[s_t.first].size() < 2) {
+      if (s_t.first != s_t.second or best_priorities[s_t.first] > 0 or
+          _sol[s_t.first].size() < 2) {
         continue;
       }
 
@@ -1045,7 +1129,8 @@ void LocalSearch<Route,
 
     // Intra Or-opt stuff
     for (const auto& s_t : s_t_pairs) {
-      if (s_t.first != s_t.second or _sol[s_t.first].size() < 4) {
+      if (s_t.first != s_t.second or best_priorities[s_t.first] > 0 or
+          _sol[s_t.first].size() < 4) {
         continue;
       }
       for (unsigned s_rank = 0; s_rank < _sol[s_t.first].size() - 1; ++s_rank) {
@@ -1108,7 +1193,8 @@ void LocalSearch<Route,
 
       // P&D relocate stuff
       for (const auto& s_t : s_t_pairs) {
-        if (s_t.first == s_t.second or _sol[s_t.first].size() == 0) {
+        if (s_t.first == s_t.second or best_priorities[s_t.first] > 0 or
+            best_priorities[s_t.second] > 0 or _sol[s_t.first].size() == 0) {
           // Don't try to put things from an empty vehicle.
           continue;
         }
@@ -1160,7 +1246,8 @@ void LocalSearch<Route,
     if (!_input.has_homogeneous_locations()) {
       // Route exchange stuff
       for (const auto& s_t : s_t_pairs) {
-        if (s_t.second <= s_t.first or
+        if (s_t.second <= s_t.first or best_priorities[s_t.first] > 0 or
+            best_priorities[s_t.second] > 0 or
             (_sol[s_t.first].size() == 0 and _sol[s_t.second].size() == 0) or
             _sol_state.bwd_skill_rank[s_t.first][s_t.second] > 0 or
             _sol_state.bwd_skill_rank[s_t.second][s_t.first] > 0) {
@@ -1183,23 +1270,36 @@ void LocalSearch<Route,
       }
     }
 
-    // Find best overall gain.
+    // Find best overall move, first checking priority increase then
+    // best gain if no priority increase is available.
+    best_priority = 0;
     best_gain = 0;
     Index best_source = 0;
     Index best_target = 0;
 
     for (unsigned s_v = 0; s_v < _nb_vehicles; ++s_v) {
-      for (unsigned t_v = 0; t_v < _nb_vehicles; ++t_v) {
-        if (best_gains[s_v][t_v] > best_gain) {
-          best_gain = best_gains[s_v][t_v];
-          best_source = s_v;
-          best_target = t_v;
+      if (best_priorities[s_v] > best_priority) {
+        best_priority = best_priorities[s_v];
+        best_gain = best_gains[s_v][s_v];
+        best_source = s_v;
+        best_target = s_v;
+      }
+    }
+
+    if (best_priority == 0) {
+      for (unsigned s_v = 0; s_v < _nb_vehicles; ++s_v) {
+        for (unsigned t_v = 0; t_v < _nb_vehicles; ++t_v) {
+          if (best_gains[s_v][t_v] > best_gain) {
+            best_gain = best_gains[s_v][t_v];
+            best_source = s_v;
+            best_target = t_v;
+          }
         }
       }
     }
 
     // Apply matching operator.
-    if (best_gain > 0) {
+    if (best_priority > 0 or best_gain > 0) {
       assert(best_ops[best_source][best_target] != nullptr);
 
       best_ops[best_source][best_target]->apply();
@@ -1209,7 +1309,7 @@ void LocalSearch<Route,
 
 #ifndef NDEBUG
       // Update route costs.
-      auto previous_cost =
+      const auto previous_cost =
         std::accumulate(update_candidates.begin(),
                         update_candidates.end(),
                         0,
@@ -1219,12 +1319,13 @@ void LocalSearch<Route,
       for (auto v_rank : update_candidates) {
         _sol_state.update_route_cost(_sol[v_rank].route, v_rank);
       }
-      auto new_cost = std::accumulate(update_candidates.begin(),
-                                      update_candidates.end(),
-                                      0,
-                                      [&](auto sum, auto c) {
-                                        return sum + _sol_state.route_costs[c];
-                                      });
+      const auto new_cost =
+        std::accumulate(update_candidates.begin(),
+                        update_candidates.end(),
+                        0,
+                        [&](auto sum, auto c) {
+                          return sum + _sol_state.route_costs[c];
+                        });
       assert(new_cost + best_gain == previous_cost);
 #endif
 
@@ -1254,16 +1355,38 @@ void LocalSearch<Route,
       s_t_pairs.clear();
       for (auto v_rank : update_candidates) {
         best_gains[v_rank].assign(_nb_vehicles, 0);
+        best_priorities[v_rank] = 0;
+        best_ops[v_rank] = std::vector<std::unique_ptr<Operator>>(_nb_vehicles);
       }
 
       for (unsigned v = 0; v < _nb_vehicles; ++v) {
         for (auto v_rank : update_candidates) {
           if (_input.vehicle_ok_with_vehicle(v, v_rank)) {
             best_gains[v][v_rank] = 0;
+            best_ops[v][v_rank] = std::unique_ptr<Operator>();
+
             s_t_pairs.emplace_back(v, v_rank);
             if (v != v_rank) {
               s_t_pairs.emplace_back(v_rank, v);
             }
+          }
+        }
+      }
+
+      for (unsigned v = 0; v < _nb_vehicles; ++v) {
+        if (best_ops[v][v] == nullptr) {
+          continue;
+        }
+        for (auto req_u : best_ops[v][v]->required_unassigned()) {
+          if (_sol_state.unassigned.find(req_u) ==
+              _sol_state.unassigned.end()) {
+            // This move should be invalidated because a required
+            // unassigned job has been added by try_job_additions in
+            // the meantime.
+            best_gains[v][v] = 0;
+            best_priorities[v] = 0;
+            best_ops[v][v] = std::unique_ptr<Operator>();
+            s_t_pairs.emplace_back(v, v);
           }
         }
       }
@@ -1272,6 +1395,7 @@ void LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1287,6 +1411,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 void LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -1380,6 +1505,7 @@ void LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1395,6 +1521,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 Gain LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -1414,31 +1541,31 @@ Gain LocalSearch<Route,
   assert(v != v_target);
 
   Gain cost = static_cast<Gain>(INFINITE_COST);
-  auto job_index = _input.jobs[_sol[v].route[r]].index();
+  const auto job_index = _input.jobs[_sol[v].route[r]].index();
 
   if (_input.vehicles[v_target].has_start()) {
-    auto start_index = _input.vehicles[v_target].start.value().index();
-    Gain start_cost = _matrix[start_index][job_index];
+    const auto start_index = _input.vehicles[v_target].start.value().index();
+    const Gain start_cost = _matrix[start_index][job_index];
     cost = std::min(cost, start_cost);
   }
   if (_input.vehicles[v_target].has_end()) {
-    auto end_index = _input.vehicles[v_target].end.value().index();
-    Gain end_cost = _matrix[job_index][end_index];
+    const auto end_index = _input.vehicles[v_target].end.value().index();
+    const Gain end_cost = _matrix[job_index][end_index];
     cost = std::min(cost, end_cost);
   }
   if (_sol[v_target].size() != 0) {
-    auto nearest_from_rank =
+    const auto nearest_from_rank =
       _sol_state.nearest_job_rank_in_routes_from[v][v_target][r];
-    auto nearest_from_index =
+    const auto nearest_from_index =
       _input.jobs[_sol[v_target].route[nearest_from_rank]].index();
-    Gain cost_from = _matrix[nearest_from_index][job_index];
+    const Gain cost_from = _matrix[nearest_from_index][job_index];
     cost = std::min(cost, cost_from);
 
-    auto nearest_to_rank =
+    const auto nearest_to_rank =
       _sol_state.nearest_job_rank_in_routes_to[v][v_target][r];
-    auto nearest_to_index =
+    const auto nearest_to_index =
       _input.jobs[_sol[v_target].route[nearest_to_rank]].index();
-    Gain Costo = _matrix[job_index][nearest_to_index];
+    const Gain Costo = _matrix[job_index][nearest_to_index];
     cost = std::min(cost, Costo);
   }
 
@@ -1446,6 +1573,7 @@ Gain LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1461,6 +1589,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 Gain LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -1490,6 +1619,7 @@ Gain LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1505,6 +1635,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 Gain LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -1538,6 +1669,7 @@ Gain LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1553,6 +1685,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 void LocalSearch<Route,
+                 UnassignedExchange,
                  Exchange,
                  CrossExchange,
                  MixedExchange,
@@ -1595,7 +1728,7 @@ void LocalSearch<Route,
     Gain best_gain = std::numeric_limits<Gain>::min();
 
     for (std::size_t r = 0; r < _sol[v].size(); ++r) {
-      auto& current_job = _input.jobs[_sol[v].route[r]];
+      const auto& current_job = _input.jobs[_sol[v].route[r]];
       if (current_job.type == JOB_TYPE::DELIVERY) {
         continue;
       }
@@ -1645,17 +1778,17 @@ void LocalSearch<Route,
   }
 
   for (const auto& r_r : routes_and_ranks) {
-    auto v = r_r.first;
-    auto r = r_r.second;
+    const auto v = r_r.first;
+    const auto r = r_r.second;
 
     _sol_state.unassigned.insert(_sol[v].route[r]);
 
-    auto& current_job = _input.jobs[_sol[v].route[r]];
+    const auto& current_job = _input.jobs[_sol[v].route[r]];
     if (current_job.type == JOB_TYPE::SINGLE) {
       _sol[v].remove(_input, r, 1);
     } else {
       assert(current_job.type == JOB_TYPE::PICKUP);
-      auto delivery_r = _sol_state.matching_delivery_rank[v][r];
+      const auto delivery_r = _sol_state.matching_delivery_rank[v][r];
       _sol_state.unassigned.insert(_sol[v].route[delivery_r]);
 
       if (delivery_r == r + 1) {
@@ -1674,6 +1807,7 @@ void LocalSearch<Route,
 }
 
 template <class Route,
+          class UnassignedExchange,
           class Exchange,
           class CrossExchange,
           class MixedExchange,
@@ -1689,6 +1823,7 @@ template <class Route,
           class PDShift,
           class RouteExchange>
 utils::SolutionIndicators LocalSearch<Route,
+                                      UnassignedExchange,
                                       Exchange,
                                       CrossExchange,
                                       MixedExchange,
@@ -1707,6 +1842,7 @@ utils::SolutionIndicators LocalSearch<Route,
 }
 
 template class LocalSearch<TWRoute,
+                           vrptw::UnassignedExchange,
                            vrptw::Exchange,
                            vrptw::CrossExchange,
                            vrptw::MixedExchange,
@@ -1723,6 +1859,7 @@ template class LocalSearch<TWRoute,
                            vrptw::RouteExchange>;
 
 template class LocalSearch<RawRoute,
+                           cvrp::UnassignedExchange,
                            cvrp::Exchange,
                            cvrp::CrossExchange,
                            cvrp::MixedExchange,