+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
+- Local search move removing a job in a route and adding an unassigned one (#324)
 - Support for string description of vehicles and tasks (#235)
+- Priority sum report in output (#390)
 
 ### Changed